requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.31.0
beautifulsoup4>=4.12.0
langchain>=0.1.0
pandas>=2.0.0
numpy>=1.24.0 